from-rag-to-deploy/pyproject.toml
[tool.poetry]
name = "from-rag-to-deploy"
version = "0.1.0"
description = ""
authors = ["Albert Gil <albert@albertgilopez.es>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain-community = "^0.2.5"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.8"
pypdf = "^4.2.0"
faiss-cpu = "^1.8.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
